# packages/api/src/openapi/statistics.yaml
openapi: 3.0.3
info:
  title: Primo Poker Statistics API
  description: API endpoints for retrieving player statistics and leaderboard data
  version: 1.0.0
servers:
  - url: https://primo-poker-server.alabamamike.workers.dev/api
    description: Production server
paths:
  /players/{playerId}/statistics:
    get:
      summary: Get player statistics
      description: Retrieve statistics for a specific player. Users can only access their own statistics unless they have admin privileges.
      tags:
        - Statistics
      security:
        - bearerAuth: []
      parameters:
        - name: playerId
          in: path
          required: true
          description: UUID of the player
          schema:
            type: string
            format: uuid
        - name: period
          in: query
          required: false
          description: Time period for statistics
          schema:
            type: string
            enum: [all_time, daily, weekly, monthly, yearly]
            default: all_time
        - name: gameType
          in: query
          required: false
          description: Type of game
          schema:
            type: string
            enum: [all, cash, tournament, sit_n_go]
            default: all
      responses:
        '200':
          description: Successfully retrieved player statistics
          headers:
            Cache-Control:
              schema:
                type: string
                example: public, max-age=300, s-maxage=300
            X-RateLimit-Limit:
              schema:
                type: integer
                example: 30
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: 25
            X-RateLimit-Reset:
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      playerId:
                        type: string
                        format: uuid
                      statistics:
                        type: array
                        items:
                          $ref: '#/components/schemas/PlayerStatistics'
                      period:
                        type: string
                      gameType:
                        type: string
        '400':
          description: Invalid player ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Cannot access other player's statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No statistics found for player
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
                description: Number of seconds until rate limit resets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /leaderboards:
    get:
      summary: Get leaderboard data
      description: Retrieve leaderboard rankings with various filters and sorting options
      tags:
        - Statistics
      parameters:
        - name: period
          in: query
          required: false
          description: Time period for leaderboard
          schema:
            type: string
            enum: [all_time, daily, weekly, monthly, yearly]
            default: all_time
        - name: gameType
          in: query
          required: false
          description: Type of game
          schema:
            type: string
            enum: [all, cash, tournament, sit_n_go]
            default: all
        - name: sortBy
          in: query
          required: false
          description: Field to sort by
          schema:
            type: string
            enum: [winnings, handsPlayed, winRate, roi]
            default: winnings
        - name: sortOrder
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: limit
          in: query
          required: false
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: playerIds
          in: query
          required: false
          description: Comma-separated list of player IDs to filter by
          schema:
            type: string
            example: "uuid1,uuid2,uuid3"
      responses:
        '200':
          description: Successfully retrieved leaderboard
          headers:
            Cache-Control:
              schema:
                type: string
                example: public, max-age=60, s-maxage=60
            X-RateLimit-Limit:
              schema:
                type: integer
                example: 30
            X-RateLimit-Remaining:
              schema:
                type: integer
                example: 25
            X-RateLimit-Reset:
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      leaderboard:
                        type: array
                        items:
                          $ref: '#/components/schemas/PlayerStatsView'
                      query:
                        type: object
                        properties:
                          period:
                            type: string
                          gameType:
                            type: string
                          sortBy:
                            type: string
                          sortOrder:
                            type: string
                          limit:
                            type: integer
                          offset:
                            type: integer
                      pagination:
                        type: object
                        properties:
                          limit:
                            type: integer
                          offset:
                            type: integer
                          hasMore:
                            type: boolean
                          nextOffset:
                            type: integer
                            nullable: true
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
                description: Number of seconds until rate limit resets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    PlayerStatistics:
      type: object
      properties:
        id:
          type: string
          format: uuid
        playerId:
          type: string
          format: uuid
        period:
          type: string
          enum: [all_time, daily, weekly, monthly, yearly]
        gameType:
          type: string
          enum: [all, cash, tournament, sit_n_go]
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
          nullable: true
        handsPlayed:
          type: integer
          minimum: 0
        handsWon:
          type: integer
          minimum: 0
        showdownsWon:
          type: integer
          minimum: 0
        showdownsSeen:
          type: integer
          minimum: 0
        totalBetAmount:
          type: number
          minimum: 0
        totalWinnings:
          type: number
          minimum: 0
        totalRakeContributed:
          type: number
          minimum: 0
        biggestPotWon:
          type: number
          minimum: 0
        vpip:
          type: number
          minimum: 0
          maximum: 100
          description: Voluntarily Put money In Pot percentage
        pfr:
          type: number
          minimum: 0
          maximum: 100
          description: Pre-Flop Raise percentage
        threeBet:
          type: number
          minimum: 0
          maximum: 100
          description: 3-bet percentage
        foldToThreeBet:
          type: number
          minimum: 0
          maximum: 100
        aggressionFactor:
          type: number
          minimum: 0
          description: (Bet + Raise) / Call ratio
        aggressionFrequency:
          type: number
          minimum: 0
          maximum: 100
          description: Percentage of non-check actions that are aggressive
        cBet:
          type: number
          minimum: 0
          maximum: 100
          description: Continuation bet percentage
        foldToCBet:
          type: number
          minimum: 0
          maximum: 100
        wtsd:
          type: number
          minimum: 0
          maximum: 100
          description: Went To ShowDown percentage
        wsd:
          type: number
          minimum: 0
          maximum: 100
          description: Won at ShowDown percentage
        sessionsPlayed:
          type: integer
          minimum: 0
        totalSessionDuration:
          type: integer
          minimum: 0
          description: Total session duration in seconds
        profitableSessions:
          type: integer
          minimum: 0
        tournamentsPlayed:
          type: integer
          minimum: 0
          nullable: true
        tournamentsWon:
          type: integer
          minimum: 0
          nullable: true
        tournamentsCashed:
          type: integer
          minimum: 0
          nullable: true
        averageFinishPosition:
          type: number
          minimum: 0
          nullable: true
        totalBuyIns:
          type: number
          minimum: 0
          nullable: true
        totalCashes:
          type: number
          minimum: 0
          nullable: true
        roi:
          type: number
          nullable: true
          description: Return on Investment percentage
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastCalculatedAt:
          type: string
          format: date-time

    PlayerStatsView:
      type: object
      properties:
        playerId:
          type: string
          format: uuid
        username:
          type: string
        lifetimeHandsPlayed:
          type: integer
          minimum: 0
        lifetimeWinnings:
          type: number
        lifetimeWinRate:
          type: number
          description: Big blinds per 100 hands
        recentHandsPlayed:
          type: integer
          minimum: 0
          description: Hands played in last 30 days
        recentWinnings:
          type: number
          description: Winnings in last 30 days
        recentWinRate:
          type: number
          description: Recent win rate (BB/100)
        overallRank:
          type: integer
          minimum: 1
          nullable: true
        profitRank:
          type: integer
          minimum: 1
          nullable: true
        volumeRank:
          type: integer
          minimum: 1
          nullable: true
        achievementsCount:
          type: integer
          minimum: 0
        lastActiveAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid query parameters
            retryAfter:
              type: integer
              description: Seconds until rate limit resets (only for 429 responses)